[GEOS-11340] Extracted `HTMLTemplateManager`.
